ARTP3 v0.5.1
fix a bug in rawData.setup when a gene was completed deleted while a
list of gene’s names are not updated accordingly. The function
filter.raw.geno is also modified when all inputed SNPs are deleted. In
that case, we should not stop the program, but need to move to the next
gene/chromosome