ARTP3 v0.8.15
give an error message when a gene is included in more than one
chromosome; fix a bug in v0.8.10-v0.8.14 after geno.files was
introducing in pathway.rawData. This bug occurs is because the
covariance matrix is sorted by SNP names and when there is only one SNP
in a chromosome, the matrix can be degenerated to a vector rather than
a 1x1 matrix. I fixed this bug by adding drop = FALSE, and also
modifying setup in function validate.setup