ARTP2 v0.9.7
major update. modify example data for function meta so that one of the
summary file does not contain a column Direction, in that case,
Direction is determined by the column BETA and a warning will be thrown
out; Modify the Direction in output of meta so that it is combined from
Direction in summary.files; add a new function
multiple.comparison.summaryData for adjusting multiple comparison when
a few pathways are tested simultaneously; several relevant functions
are added, including a c function artp2_select_genes; this function has
not been optimized for very large pathway or for a lot of pathways, and
it is very unlikely to be optimized further for this situation even in
the future due to the bottleneck of computational burden.