ARTP2 v0.9.26
This is a beta update for comparing different approaches in estimating
the association correlation matrix. To do so, a new option min.n was
added. Furthermore, two new functions, inflation and create.direction
were also added to calculate the inflation factor and create study
direction based on sample size matching. The function create.direction
has not been well tested and might be removed in the future. A few
minor updates were also made.