ARTP2 v0.9.38
major update. (1) allow for SNP without a rs number to be in pathway
definition. Those SNP should be named as 5:1234 or C5:P1234,
representing a SNP locating on chromosome 5 at 1234bp.  SNPs named by
any other format are discarded. For a SNP without a rs number, it could
be blank (or .) in a plink bim file, our package will convert it to
proper format; (2) updated read.bed and example data (rs10864787 ->
1:231017100) accordingly; (3) optimize function align.reference a
little bit.