ARTP2 v0.9.39
major update. Extend the input sel.snps so that, for a SNP with a rs
number, specifying its identity as 1:1234 or C1P1234 could also be
accepted. In previous version, a SNP with a rs number must be specified
by its rs number in sel.snps if it has a rs number in plink files.